+++ lib/src/request/remove_service.dart
@@ -1,0 +1,14 @@
+import 'package:freezed_annotation/freezed_annotation.dart';
+
+part 'remove_service.freezed.dart';
+part 'remove_service.g.dart';
+
+@freezed
+class RemoveServiceRequest with _$RemoveServiceRequest {
+  const factory RemoveServiceRequest({required String name}) =
+      _RemoveServiceRequest;
+  const RemoveServiceRequest._();
+
+  factory RemoveServiceRequest.fromJson(Map<String, dynamic> json) =>
+      _$RemoveServiceRequestFromJson(json);
+}