+++ lib/src/transport/archethic_wallet_client.dart
@@ -3,6 +3,7 @@ import 'dart:async';
 import 'dart:io';
 
 import 'package:archethic_wallet_client/archethic_wallet_client.dart';
+import 'package:archethic_wallet_client/src/request/remove_service.dart';
 import 'package:archethic_wallet_client/src/transport/common/awc_json_rpc_client.dart';
 import 'package:archethic_wallet_client/src/transport/message_channel/message_channel_desktop.dart'
     if (dart.library.js) 'package:archethic_wallet_client/src/transport/message_channel/message_channel.dart';
@@ -158,6 +159,10 @@ $transportMethodsReport
     AddServiceRequest data,
   );
 
+  Future<Result<SendTransactionResult, Failure>> removeService(
+    RemoveServiceRequest data,
+  );
+
   Future<Result<GetServicesFromKeychainResult, Failure>>
       getServicesFromKeychain();
 
@@ -171,5 +176,9 @@ $transportMethodsReport
 
   Future<Result<SignTransactionsResult, Failure>> signTransactions(
     SignTransactionRequest data,
+  );
+
+  Future<Result<SignPayloadsResult, Failure>> signPayloads(
+    SignPayloadRequest data,
   );
 }