+++ lib/src/transport/common/awc_json_rpc_client.dart
@@ -3,6 +3,7 @@ import 'dart:convert';
 
 import 'package:archethic_wallet_client/archethic_wallet_client.dart';
 import 'package:archethic_wallet_client/src/core/task.dart';
+import 'package:archethic_wallet_client/src/request/remove_service.dart';
 import 'package:json_rpc_2/json_rpc_2.dart';
 import 'package:logging/logging.dart';
 import 'package:stream_channel/stream_channel.dart';
@@ -304,6 +305,19 @@ class AWCJsonRPCClient implements ArchethicDAppClient {
       );
 
   @override
+  Future<Result<SendTransactionResult, Failure>> removeService(
+    RemoveServiceRequest data,
+  ) =>
+      Result.guard(
+        () => _send(
+          method: 'removeService',
+          params: data.toJson(),
+        ).then(
+          (result) => SendTransactionResult.fromJson(result),
+        ),
+      );
+
+  @override
   Future<Result<GetServicesFromKeychainResult, Failure>>
       getServicesFromKeychain() => Result.guard(
             () => _send(method: 'getServicesFromKeychain').then(
@@ -347,6 +361,19 @@ class AWCJsonRPCClient implements ArchethicDAppClient {
           params: data.toJson(),
         ).then(
           (result) => SignTransactionsResult.fromJson(result),
+        ),
+      );
+
+  @override
+  Future<Result<SignPayloadsResult, Failure>> signPayloads(
+    SignPayloadRequest data,
+  ) =>
+      Result.guard(
+        () => _send(
+          method: 'signPayloads',
+          params: data.toJson(),
+        ).then(
+          (result) => SignPayloadsResult.fromJson(result),
         ),
       );
 }