+++ lib/src/transport/deeplink/deeplink.dart
@@ -172,6 +172,23 @@ class DeeplinkArchethicDappClient implements ArchethicDAppClient {
       );
 
   @override
+  Future<Result<SendTransactionResult, Failure>> removeService(
+    RemoveServiceRequest data,
+  ) =>
+      _send(
+        requestEndpoint: 'remove_service',
+        replyEndpoint: 'remove_service_result',
+        params: data.toJson(),
+      ).then(
+        (result) => result.map(
+          failure: (failure) =>
+              Result.failure(Failure.fromDeeplinkRpcFailure(failure)),
+          success: (success) =>
+              Result.success(SendTransactionResult.fromJson(success)),
+        ),
+      );
+
+  @override
   Future<Result<GetServicesFromKeychainResult, Failure>>
       getServicesFromKeychain() async => _send(
             requestEndpoint: 'get_services_from_keychain',
@@ -235,6 +252,23 @@ class DeeplinkArchethicDappClient implements ArchethicDAppClient {
               Result.failure(Failure.fromDeeplinkRpcFailure(failure)),
           success: (success) =>
               Result.success(SignTransactionsResult.fromJson(success)),
+        ),
+      );
+
+  @override
+  Future<Result<SignPayloadsResult, Failure>> signPayloads(
+    SignPayloadRequest data,
+  ) =>
+      _send(
+        requestEndpoint: 'sign_payloads',
+        replyEndpoint: 'sign_payloads_result',
+        params: data.toJson(),
+      ).then(
+        (result) => result.map(
+          failure: (failure) =>
+              Result.failure(Failure.fromDeeplinkRpcFailure(failure)),
+          success: (success) =>
+              Result.success(SignPayloadsResult.fromJson(success)),
         ),
       );
 }